Round to Multiple

Returns a number rounded to the nearest specified multiple.

Usage

RoundTo(x, multiple = 1, FUN = round)

Arguments

x

is a vector of value in decibel (dB)

multiple

is a vector of time in seconds

FUN

the rounding function as character or as expression. Can be one out of trunc, ceiling, round (default) or floor.

Author

Andri Signorell andri@signorell.net

Examples

#data("P1FA")

RoundTo(x = P1FA$LAeq, multiple = 0.5)
#>    [1] 58.0 52.0 47.0 47.0 47.0 47.5 48.0 47.0 48.5 51.0 45.5 46.5 46.0 44.5
#>   [15] 46.0 48.0 46.5 45.5 46.0 46.0 46.0 45.5 46.0 46.0 47.0 51.5 53.0 50.5
#>   [29] 52.5 53.5 49.0 49.5 50.0 50.0 49.0 48.0 47.0 47.0 46.5 45.5 47.5 46.0
#>   [43] 47.0 45.0 45.5 45.0 45.0 46.0 46.0 46.0 47.0 46.5 48.0 47.5 48.5 48.5
#>   [57] 51.5 49.0 47.5 53.0 49.0 46.5 47.0 46.0 45.5 45.0 45.0 45.0 45.0 45.0
#>   [71] 44.5 44.5 45.0 44.5 44.5 44.0 44.0 44.0 44.0 44.5 44.0 44.0 44.0 45.0
#>   [85] 44.5 45.5 45.5 46.0 47.0 46.5 46.5 46.5 47.0 46.0 46.5 47.0 46.5 46.0
#>   [99] 46.5 46.0 45.5 45.5 44.5 44.5 45.0 45.0 44.5 44.5 45.5 44.5 45.0 45.0
#>  [113] 44.0 45.0 46.5 45.0 45.0 44.5 44.5 44.0 44.0 44.5 46.0 45.0 46.0 45.0
#>  [127] 45.5 45.0 44.5 45.0 45.0 45.5 47.0 46.0 46.0 46.0 45.5 46.0 46.0 45.5
#>  [141] 45.0 44.5 45.0 44.5 44.5 44.5 44.5 45.0 45.0 45.5 46.0 47.0 46.5 46.0
#>  [155] 46.5 46.0 45.5 46.0 46.5 47.0 47.5 47.5 46.5 46.5 47.0 45.5 47.0 46.0
#>  [169] 47.5 45.5 45.0 45.0 45.0 44.5 45.5 44.5 44.5 44.5 44.0 44.0 44.5 45.0
#>  [183] 46.0 47.0 47.0 47.5 48.0 47.0 48.5 47.0 46.5 47.0 48.5 47.0 47.0 46.5
#>  [197] 47.5 47.0 47.0 46.0 46.5 46.0 45.5 45.5 46.5 46.5 46.0 46.0 46.0 45.5
#>  [211] 45.5 46.5 46.0 47.0 47.0 47.0 47.5 47.5 47.5 47.0 47.0 47.0 47.5 47.5
#>  [225] 46.5 47.5 47.0 46.5 46.5 45.0 44.5 44.0 44.0 44.0 44.5 44.5 44.0 44.0
#>  [239] 44.0 45.0 46.0 45.5 45.5 44.5 44.5 44.5 44.5 44.5 44.0 44.0 55.0 45.0
#>  [253] 44.5 45.0 45.0 45.0 53.5 49.5 46.0 45.5 46.0 45.5 45.0 45.5 45.5 46.5
#>  [267] 46.0 46.5 47.5 47.5 48.5 50.5 50.0 49.5 50.5 50.0 50.0 49.5 50.0 49.0
#>  [281] 47.0 46.5 46.0 45.5 45.0 45.0 45.0 44.5 44.5 45.0 45.0 46.0 45.5 46.5
#>  [295] 45.5 46.0 48.0 47.5 46.0 48.5 49.0 46.5 45.5 46.0 47.0 47.0 47.5 47.5
#>  [309] 47.5 48.0 47.0 49.0 48.5 48.0 46.5 45.5 45.5 45.5 47.0 46.5 46.5 47.0
#>  [323] 48.5 48.0 47.5 49.0 47.5 46.5 46.5 49.5 52.5 51.5 47.5 47.0 46.5 50.0
#>  [337] 50.0 47.5 47.0 50.0 48.5 47.5 49.5 48.0 54.5 54.5 50.5 48.0 48.0 48.5
#>  [351] 46.0 45.0 48.5 49.0 53.5 49.5 50.0 52.0 54.0 51.5 50.5 50.0 51.0 56.0
#>  [365] 55.5 55.0 52.5 52.0 51.0 50.0 49.5 50.0 48.5 50.0 49.5 49.5 50.0 51.0
#>  [379] 51.0 48.0 49.0 46.5 45.5 45.0 44.5 45.0 45.5 46.0 45.5 45.5 45.5 45.0
#>  [393] 45.5 45.0 46.0 48.5 48.0 47.0 45.5 45.5 45.0 45.0 46.0 45.5 45.5 46.0
#>  [407] 46.0 46.0 45.5 44.5 44.5 44.5 44.5 45.0 45.0 45.5 45.5 48.0 48.0 48.5
#>  [421] 49.0 51.5 49.5 48.5 50.0 47.5 46.0 46.0 47.5 50.5 54.0 56.0 50.0 45.5
#>  [435] 45.5 45.5 46.0 46.5 48.5 52.5 54.0 57.0 55.5 48.0 45.5 45.5 46.5 47.0
#>  [449] 47.5 52.0 52.5 50.5 53.0 51.5 49.0 50.0 49.5 48.0 46.5 47.5 46.5 46.0
#>  [463] 45.5 45.0 44.5 44.5 44.5 45.0 44.5 44.5 44.5 44.5 44.5 44.0 44.5 45.0
#>  [477] 45.5 45.5 46.0 45.0 45.0 45.0 45.5 46.0 45.5 46.0 45.5 45.5 45.5 45.5
#>  [491] 46.0 45.5 46.0 44.5 44.0 44.5 44.5 45.0 45.5 50.5 56.0 52.5 47.5 45.0
#>  [505] 44.5 44.5 44.5 44.5 45.0 45.0 46.5 46.5 47.0 47.0 46.5 47.0 48.5 47.5
#>  [519] 48.5 48.0 48.5 48.5 47.5 47.0 47.0 46.5 46.5 46.0 45.5 46.0 45.5 45.5
#>  [533] 45.0 44.5 44.5 44.5 45.5 46.0 45.5 45.5 45.0 45.0 44.5 44.5 44.0 44.5
#>  [547] 45.0 45.0 45.0 45.5 45.5 46.5 48.5 48.5 48.5 47.5 47.0 47.0 45.5 45.0
#>  [561] 46.5 46.5 47.0 45.5 46.0 45.0 44.5 45.0 44.5 45.0 45.0 45.0 46.0 47.5
#>  [575] 45.5 45.5 44.5 44.5 45.0 44.5 44.5 45.0 49.5 47.5 46.0 47.0 46.0 47.5
#>  [589] 47.0 46.5 46.5 45.0 45.5 46.0 47.0 49.0 50.5 45.5 44.5 44.5 44.5 44.5
#>  [603] 46.5 45.0 45.5 47.0 46.0 46.0 45.5 44.5 44.5 44.5 44.5 44.5 45.0 45.0
#>  [617] 45.5 45.5 45.5 46.5 46.5 46.5 47.0 47.0 47.5 46.5 46.5 46.5 46.5 45.5
#>  [631] 46.0 45.0 44.5 44.0 44.5 44.0 44.0 44.0 44.5 44.5 44.5 46.0 45.5 46.0
#>  [645] 48.5 47.0 47.5 46.0 45.5 45.0 46.0 46.5 46.5 48.0 47.0 48.0 48.0 48.0
#>  [659] 47.0 46.0 46.0 45.5 46.5 47.0 45.0 45.5 45.5 48.0 47.0 46.5 46.0 46.0
#>  [673] 45.5 45.5 45.5 47.5 48.5 47.0 47.5 47.0 46.5 46.0 47.0 46.5 47.5 48.5
#>  [687] 46.0 46.0 46.5 45.5 45.5 44.5 44.5 44.5 44.5 45.0 44.5 45.0 46.0 45.5
#>  [701] 45.5 45.0 45.5 45.5 45.0 45.5 46.0 46.0 46.0 46.0 46.0 46.5 47.0 46.5
#>  [715] 46.5 46.0 45.5 46.5 46.0 46.5 45.5 45.5 45.5 45.0 45.0 45.5 46.5 46.5
#>  [729] 46.5 47.0 46.5 46.0 46.0 45.0 45.0 44.5 44.5 44.5 45.0 45.0 45.5 45.5
#>  [743] 45.5 46.5 47.5 49.5 50.0 48.5 47.0 48.0 47.0 48.0 49.0 48.0 47.5 47.5
#>  [757] 48.5 48.0 46.0 45.5 45.5 45.5 46.0 47.0 46.5 47.0 46.0 45.0 45.0 44.5
#>  [771] 44.5 44.5 44.5 45.0 45.0 46.0 46.5 47.0 47.5 47.5 47.5 47.5 47.5 48.5
#>  [785] 53.0 52.0 52.0 52.0 57.5 54.0 52.5 56.5 54.5 49.0 47.5 46.5 46.0 45.5
#>  [799] 45.0 45.0 45.0 44.5 44.5 44.5 45.0 44.5 45.0 45.5 45.0 45.5 46.5 46.0
#>  [813] 46.0 47.0 47.0 46.0 46.0 45.5 45.0 45.5 44.5 44.5 44.5 44.0 45.0 45.0
#>  [827] 45.0 44.0 44.5 44.5 44.5 45.0 45.5 46.5 46.0 45.5 46.0 45.5 46.0 46.5
#>  [841] 46.5 46.0 45.5 45.0 44.5 44.5 44.5 44.5 45.0 45.0 45.5 46.5 46.0 46.0
#>  [855] 45.5 46.0 47.0 47.5 48.0 47.0 47.0 46.0 46.0 45.5 45.5 46.0 46.5 46.5
#>  [869] 50.5 48.0 48.5 48.5 47.0 46.5 46.0 46.0 45.5 45.0 45.0 45.0 45.0 45.5
#>  [883] 46.0 47.5 48.0 48.0 47.0 46.5 46.0 44.5 44.5 44.5 44.5 45.0 45.0 44.5
#>  [897] 45.5 45.0 45.0 45.0 44.5 44.5 44.5 44.5 44.5 44.5 45.0 45.5 47.0 47.5
#>  [911] 47.0 47.0 47.5 47.0 47.0 45.0 45.0 44.5 44.5 44.5 44.5 44.5 44.0 44.0
#>  [925] 44.0 44.5 44.5 44.5 44.0 44.5 44.5 44.0 44.0 44.0 44.5 44.5 44.5 44.5
#>  [939] 44.5 44.0 44.0 44.5 45.0 45.0 45.5 45.5 46.0 45.5 46.0 46.5 47.0 47.5
#>  [953] 48.5 47.5 48.0 47.0 46.0 45.5 45.0 44.5 44.0 44.0 44.5 44.5 44.5 44.5
#>  [967] 44.5 45.0 45.5 45.0 45.5 45.5 46.5 47.0 45.0 45.0 45.5 45.0 45.5 46.0
#>  [981] 46.5 47.5 47.0 47.0 45.5 46.0 45.0 44.5 44.5 44.5 44.5 44.5 44.0 44.0
#>  [995] 44.0 44.5 44.5 44.5 45.0 45.0 45.0 46.0 46.5 46.0 47.0 47.0 45.5 46.0
#> [1009] 45.0 44.5 44.5 44.5 44.5 44.0 44.0 44.0 44.5 44.5 44.5 44.5 45.0 45.0
#> [1023] 45.5 45.5 46.0 45.0 45.5 46.0 46.0 46.5 46.5 46.0 46.0 45.5 44.5 44.5
#> [1037] 44.5 45.0 45.0 46.0 47.5 47.5 48.0 49.0 49.5 49.0 51.5 54.0 52.5 52.0
#> [1051] 50.0 50.0 48.0 45.5 46.5 46.0 46.5 46.5 46.5 47.0 48.5 49.0 48.0 47.0
#> [1065] 48.0 48.0 47.0 47.0 47.0 46.5 47.0 46.5 46.5 46.0 46.0 46.0 45.5 45.5
#> [1079] 46.5 46.5 46.0 46.0 46.5 46.0 46.5 48.0 46.0 46.0 45.0 45.0 45.5 44.5
#> [1093] 44.5 44.0 44.5 44.5 44.5 45.5 46.5 44.0 44.5 44.0 44.5 45.5 45.5 46.5
#> [1107] 47.0 46.5 47.0 47.5 46.0 44.5 44.5 44.5 45.0 45.5 46.5 48.5 49.5 48.0
#> [1121] 48.5 49.0 47.0 47.5 47.5 47.5 47.0 46.5 46.0 47.0 46.5 47.0 46.5 45.5
#> [1135] 45.5 46.0 47.0 46.5 48.0 47.0 45.5 45.5 46.0 46.0 46.0 47.0 47.5 47.5
#> [1149] 49.0 49.0 48.0 47.5 47.5 47.0 47.5 46.5 46.0 45.5 46.5 46.0 45.5 46.0
#> [1163] 44.5 44.0 44.0 44.0 44.0 44.0 44.0 45.5 46.0 45.5 45.5 45.0 45.0 44.5
#> [1177] 44.5 44.0 44.0 44.0 44.0 44.0 44.0 44.5 45.0 45.0 46.5 46.0 46.0 46.0
#> [1191] 45.5 45.0 46.0 46.0 48.5 51.0 49.0 47.5 45.5 44.5 44.5 44.0 44.0 44.5
#> [1205] 44.5 45.5 45.5 45.0 45.0 45.5 46.0 46.0 46.0 46.5 47.5 49.5 48.0 47.5
#> [1219] 48.5 49.0 46.5 46.0 46.0 46.0 46.0 46.5 47.5 46.0 47.0 48.0 46.5 47.5
#> [1233] 47.5 48.5 48.0 52.5 55.5 57.5 54.5 50.5 47.5 46.5 45.0 46.0 45.0 45.5
#> [1247] 47.5 44.5 45.5 48.5 47.5 45.0 44.5 44.5 44.5 44.5 44.0 44.0 44.0 44.0
#> [1261] 44.0 44.0 44.5 45.0 45.0 45.0 46.5 46.0 46.5 46.0 45.5 45.5 45.0 45.0
#> [1275] 45.0 44.5 44.0 44.0 44.5 44.0 44.0 44.0 44.0 44.0 44.0 44.0 44.0 44.0
#> [1289] 44.0 44.0 45.5 44.5 44.0 44.0 44.0 44.0 44.5 44.0 44.5 45.5 44.5 44.5
#> [1303] 44.5 44.5 45.0 44.5 45.5 46.0 46.0 46.5 46.0 45.5 45.0 45.0 44.5 44.5
#> [1317] 44.5 44.5 44.5 46.0 47.0 47.5 46.5 47.0 46.5 45.5 46.0 46.0 45.5 45.5
#> [1331] 46.0 46.0 46.5 47.0 47.5 46.5 46.5 46.5 46.0 45.5 46.5 46.0 47.0 46.5
#> [1345] 46.0 46.0 47.0 46.5 47.0 49.5 49.0 48.5 49.0 48.5 48.0 47.0 47.0 46.5
#> [1359] 45.5 45.5 45.5 46.0 46.0 45.5 45.5 46.0 46.0 47.5 49.0 47.5 49.5 51.5
#> [1373] 51.0 47.5 49.0 50.5 49.5 48.0 48.0 47.5 46.5 47.0 45.5 46.0 46.0 45.5
#> [1387] 45.5 45.5 45.0 44.5 44.0 44.0 44.0 44.0 44.5 44.5 45.5 46.0 47.0 47.5
#> [1401] 55.0 47.5 49.5 46.5 44.5 44.0 54.5 53.0 44.0 49.5 44.0 44.0 51.5 46.5
#> [1415] 53.0 51.0 46.0 46.5 46.0 55.5 47.5 53.5 47.5 48.0 55.5 54.0 50.0 55.5
#> [1429] 46.5 46.0 45.0 44.5 44.5 45.0 44.5 45.0 45.5 46.0 45.5 46.0 46.5 46.5
#> [1443] 48.0 47.5 48.0 47.0 47.0 47.0 46.0 46.5 47.0 47.0 47.5 48.0 47.5 47.5
#> [1457] 46.0 46.0 45.0 45.0 44.5 44.5 44.5 44.5 45.0 44.5 44.5 44.5 45.5 46.0
#> [1471] 45.5 46.0 47.0 47.0 48.0 50.0 48.5 47.5 50.0 49.0 50.5 50.5 50.5 50.0
#> [1485] 51.0 54.5 57.0 58.5 58.5 58.0 56.0 55.5 56.0 58.5 61.0 60.0 54.5 53.0
#> [1499] 54.0 55.0 55.0 56.0 56.0 55.0 53.0 50.5 50.5 48.5 49.0 49.5 47.0 46.5
#> [1513] 46.0 45.5 45.0 45.0 45.0 44.5 44.5 44.5 45.0 45.5 45.5 45.5 46.0 46.5
#> [1527] 46.5 47.0 46.0 46.5 47.5 47.0 48.0 45.0 44.0 45.5 45.0 44.5 44.5 45.5
#> [1541] 45.5 45.0 45.0 44.5 45.5 45.0 44.5 44.0 44.0 44.0 44.5 44.0 44.0 44.0
#> [1555] 44.0 44.0 44.0 44.0 44.5 45.0 44.5 45.0 45.0 45.0 45.0 45.5 45.5 45.5
#> [1569] 46.0 47.0 45.5 46.5 45.5 45.5 62.0 53.5 45.5 45.5 46.0 45.0 46.5 46.5
#> [1583] 46.5 47.0 47.0 45.0 44.5 44.5 45.0 44.5 45.0 45.0 46.0 46.0 46.0 45.5
#> [1597] 45.5 45.0 45.0 44.5 44.5 44.5 44.5 45.0 45.0 45.5 55.5 48.0 47.5 48.0
#> [1611] 47.0 49.0 51.0 50.0 51.0 51.0 49.5 50.0 50.0 49.5 48.0 46.0 46.5 46.0
#> [1625] 45.5 46.0